working fork template
